requirements.txt
-i http://pypi.doubanio.com/simple/
Django~=1.11
factory-boy~=2.12.0
Faker~=2.0.3
-e .
